+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.0.12
+* Added Models: ChatInput, ChatJoin, ChatMessage, ChatRoom, UserModel
+* Added Chat Room Widgets: ChatRoomScreen, ChatBubble, ChatRoomDoc, ChatRoomInputBox
+* Updated ChatService methods: join(), showChatRoomScreen(), sendMessage()
+* Initial codes for invitation.
+
+## 0.0.11
+* Bump to 11
+
+## 0.0.10
+* Fix. Mirror firestore user data to database.
+
+## 0.0.9
+* Bump to 9
 
 ## 0.0.8
 * Bump to 8